requirements.txt
# Added during Protocol 0 Initialization
python-binance==1.0.19
python-dotenv==1.0.1
rich==13.7.1
click==8.1.7
python-dotenv==1.2.2
